mpr_int/requirements.txt
numpy
pandas
scipy
python-socketio
werkzeug
openpyxl
requests
